Refactored product analytics models files
